Use cat instead of lena. Do not use data from vispy, because 0.1.0 does not have cube and cat yet.

+++ cube_v3.py
@@ -5,7 +5,8 @@ Show spinning cube using VBO's, and transforms.
 """
 
 import numpy as np
-from vispy import app, gl, oogl, io
+from vispy import app, gl, oogl
+import vispy_io as io  # Because vispy 0.1.0 lacks some data files
 from transforms import perspective, translate, rotate
 
 
@@ -41,7 +42,7 @@ void main()
 positions, faces, normals, texcoords = io.read_mesh('cube.obj')
 colors = np.random.uniform(0,1,positions.shape).astype('float32')
 
-faces_buffer = oogl.ElementBuffer(faces)
+faces_buffer = oogl.ElementBuffer(faces.astype(np.uint16))
 
 
 class Canvas(app.Canvas):
